// packages/sdk/src/contracts/generated.ts
export type StructFromName<T extends string | undefined> = T extends keyof GeneratedStructMap ? GeneratedStructMap[T] : undefined;



export type StructWireType = 'uint32' | 'string' | 'uint64' | 'sint64' | 'array' | 'sint8' | 'asset' | 'unknown' | 'sint32' | 'uint8' | 'uint16' | 'sint16' | 'struct' | never;

export interface StructLayoutField {
  name: string;
  wireType: StructWireType;
  structName?: string;
  length?: number;
  elementType?: StructWireType;
  elementStructName?: string;
}